No longer require XCom pickling for 2.2.5
In the current implementation, we pass a "DatabricksMetadata" object via Xcom that contains basic metadata like run id and conn id. When supporting earlier versions of Airflow, this would mean that users have to turn on xcom pickling (which would be a security risk). To avoid this risk we simply return a json dictionary that we can then manually deserialize as needed.