No longer require XCom pickling for 2.2.5 (#15)
Previously, we passed a "DatabricksMetadata" object via Xcom that contains basic metadata like run id and conn id between `TaskGroup` and inner notebook `Operators`. The plugin also used this.
When supporting earlier versions of Airflow, this would mean that users have to turn on Xcom pickling (which would be a security risk). To avoid this risk, we return a JSON dictionary that we can manually deserialize as needed.