Fix copying dependencies from task groups to notebooks (with inner task groups) (#47)
Before this change, `DatabricksWorkflowTaskGroup` (`astro-provider-databricks==0.1.3`) did not pass Python dependencies to inner Databricks tasks if they were inside intermediate Airflow Task Groups for Airflow 2.2.4. For newer Airflow versions, it duplicated the dependencies proportionally to the amount of nested `TaskGroup`s.

In the example, DAG introduced:
- notebooks 1, 2, 3 and 4 need the `simplejson` dependency
- the `simplejson` dependency is only set at the `DatabricksWorkflowTaskGroup` level (`example_task_group`)
- notebooks 2 and 3 are inside an inner Airflow `TaskGroup` (`my_task_group`)

<img width="642" alt="Screenshot 2023-06-14 at 16 52 59" src="https://github.com/astronomer/astro-provider-databricks/assets/272048/f3c69dcb-d984-44d7-a3eb-7954c833c65c">

We confirmed the Databricks cluster does not have `simplejson` as a dependency:
<img width="869" alt="Screenshot 2023-06-14 at 16 45 10" src="https://github.com/astronomer/astro-provider-databricks/assets/272048/8d72e0ea-d232-4b42-9097-3227e0041788">

And we also confirmed the notebooks failed since the cluster does not have this dependency:
<img width="1032" alt="Screenshot 2023-06-14 at 16 48 07" src="https://github.com/astronomer/astro-provider-databricks/assets/272048/206bb0fa-8a42-40fd-af43-4bc747ab3fd5">

All the tasks from this Databricks Workflow have the `simplejson` dependency in Databricks, both the ones inside the inner task group and the ones outside. Examples:

<img width="850" alt="Screenshot 2023-06-14 at 16 49 41" src="https://github.com/astronomer/astro-provider-databricks/assets/272048/10f010b7-821b-41d2-9cf6-a9ca755eaf14">

<img width="838" alt="Screenshot 2023-06-14 at 16 49 55" src="https://github.com/astronomer/astro-provider-databricks/assets/272048/9fa403d3-6937-470c-b7f0-7e45518ad431">